SDKのMissingReferenceExceptionの修正方法を調査したが見つからなかった
